test: merge test_problem_finder.py files
